test: ensure integration between RegisterUserUseCase and User.register()
